Add UI test to ensure Case cannot be finalised if there is an open query
