+++ ui_tests/caseworker/features/give_advice.feature
@@ -526,3 +526,44 @@ Feature: I want to record my user advice and any comments and conditions relatin
     Then I see the "under the Military End-Use Control of Article 4(2) of retained Council Regulation (EC) No 428/2009" text on the document preview
     When I click continue
     Then I see "READY TO SEND" inform letter status in decision documents
+
+
+  Scenario: LU cannot finalise a case if there is an open query
+    Given I sign in to SSO or am signed into SSO
+    And I create standard application or standard application has been previously created
+    And I prepare the application for final review
+    When I go to my profile page
+    And I change my team to "Licensing Unit" and default queue to "Licensing Unit Post-circulation Cases to Finalise"
+    And I go to my case list
+    And I click the application previously created
+    And I assign myself to the case
+    And I click the queries tab
+    And I click "Send a query to the exporter"
+    And I enter "Provide data sheet for line item1" as the query
+    And I click send
+    Then I see "Provide data sheet for line item1" as the query under open queries
+    When I click the recommendations and decision tab
+    And I click "Review and combine"
+    And I enter "reason for approving" as the reasons for approving
+    And I enter "licence condition" as the licence condition
+    And I click submit recommendation
+    Then I see "reason for approving" as the reasons for approving
+    And I see "licence condition" as the licence condition
+    And I see countersign required warning message
+    When I click move case forward
+    And I go to my profile page
+    And I change my team to "Licensing Unit" and default queue to "Licensing manager countersigning"
+    And I go to my case list
+    And I click the application previously created under open queries tab
+    And I click the recommendations and decision tab
+    And I click "Review and countersign"
+    And I agree with outcome and provide "licensing manager approved" as countersign comments
+    And I click submit recommendation
+    Then I see "licensing manager approved" as countersign comments
+    When I click move case forward
+    And I go to my profile page
+    And I change my team to "Licensing Unit" and default queue to "Licensing Unit Post-circulation Cases to Finalise"
+    And I go to my case list
+    And I click the application previously created under open queries tab
+    And I click the recommendations and decision tab
+    Then I see warning that case cannot be finalised due to a query that needs to be closed
